test(dev-core,dev-init): dedup detectGitHubRepo suite into shared factory (#218)
Extract the byte-identical gh_project_id/detectGitHubRepo describe blocks
into plugins/shared/__tests__/detect-github-repo.suite.ts, parameterized by
a per-file dynamic importer + injected detectGitHubRepo so file-scoped
vi.mock hoisting and relative import resolution stay correct. Both config
test files now call the factory and remain byte-identical.
